feat: passing all passkey js code to controller
