+++ config/importmap.rb
@@ -26,7 +26,6 @@ pin "@material/menu-surface/component", to: "https://ga.jspm.io/npm:@material/me
 pin "@material/menu-surface/constants", to: "https://ga.jspm.io/npm:@material/menu-surface@14.0.0/constants.js"
 pin "@material/menu-surface/foundation", to: "https://ga.jspm.io/npm:@material/menu-surface@14.0.0/foundation.js"
 pin "@github/webauthn-json", to: "https://ga.jspm.io/npm:@github/webauthn-json@2.1.1/dist/esm/webauthn-json.js"
-pin "credential", to: "credential.js"
 pin "stub_credentials" if Rails.env.test?
 
 # sinon
